services/ingest/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
imapclient>=2.3.1
apscheduler>=3.10.4
httpx>=0.27.0
fastapi>=0.111.0
uvicorn>=0.29.0
